+++ changelog.md
@@ -9,3 +9,10 @@
 # 1.0.3
 
 -   Change Pandoc command line parameter `--reference-docx` to `--reference-doc`.
+
+
+# 1.0.4
+
+-   Add logs.
+-   Update for Foliant 1.0.4: Pass logger to spinner.
+-   Require Foliant 1.0.4.